pyproject.toml
[project]
name = "agentic-art"
version = "0.1.0"
description = "AgenticART - Mobile Security Automation Framework"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "GitSolved" }
]
keywords = ["penetration-testing", "android", "llm", "security", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "streamlit>=1.30.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "chromadb>=0.4.0",
    # Security Tools
    "frida-tools>=12.0.0",
    "objection>=1.11.0",
    "androguard>=3.4.0",
    "mitmproxy>=10.0.0",
    "drozer>=2.4.4",
    "pwntools>=4.11.0",
    "ropper>=1.13.0",
    "cryptography>=41.0.0",
    "pyopenssl>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]
# Minimal install without security tools (for CI/testing only)
minimal = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "streamlit>=1.30.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "chromadb>=0.4.0",
]

[project.scripts]
agentic-art = "webapp.app:main"

[project.urls]
Homepage = "https://github.com/GitSolved/AgenticART"
Documentation = "https://github.com/GitSolved/AgenticART#readme"
Repository = "https://github.com/GitSolved/AgenticART"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "core", "webapp"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
exclude = ["dojo_finetune_package/"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true
disable_error_code = ["var-annotated", "attr-defined", "union-attr", "call-overload", "arg-type", "assignment"]
exclude = [
    "scripts/generated/",
    "output/",
    "tests/prove_feedback_loop\\.py",
    "experiments/",
    "dojo_finetune_package/",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=agent --cov=core --cov-report=term-missing"
